fix(목표예산관리) : 수정 - BudgetGoalRequestDTO - Controller,Servie - Dashboard

+++ src/main/java/com/example/fin_mangement/entity/BudgetGoal.java
@@ -22,7 +22,7 @@ public class BudgetGoal {
     private String username;
 
     @Column(nullable = false)
-    private double targetAmount;
+    private BigDecimal  targetAmount;
 
     @Column(nullable = false)
     private LocalDate startDate;
